Polishing up the initial RiveScript-Python Distro
Add more functions to the rivescript.py class to allow the programmer to
set language handlers, objects, variables etc. from the Python code. Add
a JavaScript object example to the eg folder.

Fleshed out the README.md file. Added syntax checking support and the
strict mode.

The rivescript.py runs as a stand-alone program, and is compatible with
the `rivescript` command that comes with the Perl distribution (JSON
mode works with it too).